Networking improvements. Darts can be enabled in the code which I might make publicly accessable one day.

+++ source/net.h
@@ -138,7 +138,7 @@ void CorrectPrediction(long actualfifoplc);
 //TENSW: safe packet senders
 void netsendpacket(int ind, char *buf, int len);
 void netbroadcastpacket(char *buf, int len);
-short netgetpacket(long *ind, char *buf);
+long netgetpacket(long *ind, char *buf);
 
 
 enum MultiGameTypes
